fix: move SessionMetadata to models
